+++ zarrs_metadata/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `v3::group::{ConsolidatedMetadata,ConsolidatedMetadataMetadata,ConsolidatedMetadataKind}`
+- Add `GroupMetadataV3::consolidated_metadata` field
+- Add `GroupMetadataV3::with_consolidated_metadata` field
+
 ### Changed
 - **Breaking**: Rename `DataTypeMetadataV3::Binary` to `Bytes` for compatibility with `zarr-python`
 - **Breaking**: Revise `PcodecCodecConfiguration` to match `numcodecs`: